Move output file content to preinstaller file
